correct confusing action names in kanagrams shortcuts dialog
Kanagram has KStandardAction::prior (Qt::Key_PageUp) and next (Qt::Key_PageDown)
for the actions to switch to the previous/next Vocabulary.
But due to using KStandardAction::prior() and KStandardAction::next()
these actions are labelled 'Previous Page' and 'Next Page' in the shortcuts dialog.
Renamed to 'Previous Vocabulary' and 'Next Vocabulary' to make it clear,
which action are handled with these shortcuts.

REVIEW:101496