Eliminate the kdeui dependency inside the game engine
REVIEW: 103706

Replace the KMessageBox usage with kDebug for now so that the engine does not
depend on the kdeui module anymore. This is an important step for reducing the
dependencies with small price. It is especially important on mobile platforms,
like Harmattan on N9/N950 because kdeui would not be used there for the Ui
anyway.

The alternative solution might be later to store the error message in a member
variable, or "return" it somehow either with a return value or emitting a signal
with the relevant error string.