Modularize the whole kanagram tree into engine, desktop and harmattan for now
1) All the relevant CMakeLists.txt files were refactored according to their new
tasks inside the relevant subfolders

2) FindKanagramEngine.cmake was added for find the functionality (library)
easily sinde the frontend codes.

3) There was a kanagramharmattansettings.kcfgc addition because we cannot go
only for KCoreConfigSkeleton usage because  the KConfigDialog only accepts
KConfigSkeleton for the time being. It is not even fixable inside the KDE 4.X
series, just only for frameworks.

4) We now use the VISIBILITY option with the KConfig XT technology in order to
get the generated KanagramSettings class exported. Otherwise, it would not be
available for the frontends.

5) Messages.sh was also moved to the desktop frontend because that is now
available only. Nevertheless, it is better to keep a separate Messages.sh file
for each frontend to avoid the complexity in the future.

6) A typical "kanagram_engine_export.h" file was added for defining the export
and import symbols so that they can be used inside the game engine itself.

7) The previous kanagram.desktop file is obviously now being moved to the
desktop frontend folder since the Harmattan and other players are gonna have
such a file with distinct contents.

There are also other minor changes, but the codebase is now feasible for getting
used with other frontends easily. There are still tasks left, like making a
proper versioning for the shared library and so on, but it is better to apply
such patches in different commits since this patch is already very huge itself.