Use .at() instead of .value() since it is slightly faster
It is not a risk to use because according to the bound random index number
generation, the value will be with the bound.

Quote from the Qt documentation:
"If the index i is out of bounds, the function returns a default-constructed
value. If you are certain that the index is going to be within bounds, you can
use at() instead, which is slightly faster."