Nuke the own selection model (derivated from QStringListModel)
There is no need for such a thing anymore after using the SelectionDialog from
the master HEAD version of the git repository. The only need previously for that
was the role name change since the Harmattan component only accepted "name" role
name which is not the case for the "display role" in case of C++ models, like
this QStringListModel.

Please see the following commit for further details:
https://qt.gitorious.org/qt-components/qt-components/commit/cbeb20d68c9b792353e301363df9719f5a5e8f26