Set the orientation lock to landscape for the MainPage
Unfortunately, a bit longer words than "usual" are not visibile in Portrait
mode, thus it somewhat ruins the UX. The main page is now being locked to
Landscape, thereby. It is a similar concept to the Angry Birds or NFS ones.

The Settings page will still be auto oriented since the UX is not that bad in
there. The main page can be changed later wrt the locking if there are new UI
layout ideas.