Establish distinct SoundEffect components for each different "wav" files
Unfortunately as of now, the Qt Mobility Audio stack has a few cumbersome, but
known bugs. It involves the thing if you would like to play, and set the source
beforehand, it crashes with SoundEffect, and misbehavior with the Audio QML
Element, too. The current workaround is establish distinct components for each
sound effect. As for kanagram, it is not a huge problem because we use only
three difference SoundEffect elements.