Do not push the whole qml file, just according to the id since it is exposed
There is no need for loading the MainSettingsPage.qml file anymore from the
resource system since it is properly instantiated inside the main.qml and
available due to its "id". Yet another important bit is that it makes the
application simpler to test on Harmattan since a simple "qrc:/main.qml" ->
"main.qml" change is enough inside the main.cpp file. It has also been neccesary
so far to change the "qrc:/MainSettingsPage.qml" to "MainSettingsPage" manually.
It is not mandatory anymore.