Use Column for the timer settings so different rows for better UX
The space is very small especially in Portrait mode for the settings. The slider
might represent a bit more value than available for the relevant step size in
Portrait mode. It is essentially the same concept as for the Brightness slider
and setting. Even the word "Brightness" fills in 30-40% of the actual row, and
that is not even a long word enough in the practice. It must be at least that
font size though to be readable enough.