Link against "kdeclarative" without any variables for now
We do not have yet a find module for KDeclarative, thus it is now being done in
a plain manner, but needs to be fixed later. FindKDeclarative.cmake has
apparently been pushed against kde-workspace recently which is not really the
proper place for that ... It needs to be updated when kdelibs gets updated for
Harmattan which is probably not any soon.. Anyway, it works.