Get rid of the imported platform icon usage since apparently, image can handle
It is not really prettily documented, but it is possible to use something like
this in the QML/Qt-Components code:
iconSource: "image://theme/icon-l-user-guide-main-view";

It is a working example for a ToolIcon, but since that qt component is just
wrapping the Image, it is possible to use the "source" property for a plain
Image. Unfortunately it is not documented, but widely used in code bases and a
nice feature.