Load the words according to the system language initially
The words are loaded in English, if there is nothing set yet. This is suboptimal
for a non-English speaker since the user needs to change the language manually
at least once, even if the information can be get from the system language. If
the words are not translated (kvtml files) in the system language, English is
used as a fallback.