Add anagram feature.
Now words that are correct in the language of the document and anagrams
of the original word are also counted as a correct entry. For example if
the word is three and the user enters ether the word is counted as correct.
Developed as Summer of KDE by Ajay Singh <ajaysimple91@gmail.com>