Move qml files and icons/images into a resource.
Since QDir::locate doesn't work the same on all platforms (MacOS
and Windows fail hard here) bundle the qml files and icons into a
resource included in kanagram binary.
TODO: Need to also move the sounds into the resource or find
a better way to find them if they need to be outside of kanagram
to be playable.