Main: Prevent playerBox from being offset on resize
While I was at it, I also made the size of the `currentPlayerText` bigger. I opted for the larger \`screen.width / 68.5\` option as I thought this looked better personally, but open to suggestions. The default is not unbearable but it looks oddly out of place, in my opinion. :slight_smile: