Main: Fix moreOptionsButton being cutoff on resize
Currently, if the window is resized while the `moreOptionsButton` is visible, it does not get resized when the window is resized. This results in incorrect sizing and spacing for the icons and an unnatural look, as if the `moreOptionsButton` is hidden and shown again when the window is maximized, it will use the correct size.

The cause of this problem is that the `width` for the `moreOptionsButton` is set when it is toggled to be visible, in `toggleMoreOptions()` where it is set to `blackboard.width / 2.5`. However this does not dynamically update, meaning it is set to the `blackboard.width` at the time the button is clicked